% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{autosync_editor-shiny}
\alias{autosync_editor-shiny}
\alias{autosync_editor_output}
\alias{render_autosync_editor}
\title{Shiny bindings for autosync_editor}
\usage{
autosync_editor_output(outputId, width = "100\%", height = "400px")

render_autosync_editor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Widget dimensions (must be valid CSS unit or a number
which will be coerced to a string and have "px" appended).}

\item{expr}{An expression that generates an autosync_editor widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Logical, whether \code{expr} is a quoted expression.}
}
\value{
\code{autosync_editor_output()} returns a Shiny output element.
\code{render_autosync_editor()} returns a Shiny render function.
}
\description{
Output and render functions for using autosync_editor within Shiny
applications and interactive R Markdown documents.
}
