% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{check_auth_timeout}
\alias{check_auth_timeout}
\title{Check if a pending connection has timed out}
\usage{
check_auth_timeout(server, temp_id)
}
\arguments{
\item{server}{Server state environment.}

\item{temp_id}{Temporary connection ID.}
}
\description{
Called by later::later() to check if a client authenticated in time.
}
\keyword{internal}
