% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{discover_jwks_uri}
\alias{discover_jwks_uri}
\title{Discover the JWKS URI for an OIDC issuer}
\usage{
discover_jwks_uri(issuer)
}
\arguments{
\item{issuer}{The OIDC issuer URL.}
}
\value{
Character string, the JWKS URI.
}
\description{
Discover the JWKS URI for an OIDC issuer
}
\keyword{internal}
