% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{list_documents}
\alias{list_documents}
\title{List all document IDs}
\usage{
list_documents(server)
}
\arguments{
\item{server}{An autosync_server object.}
}
\value{
Character vector of document IDs.
}
\description{
Returns the IDs of all documents currently loaded in the server.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
server <- sync_server()
create_document(server)
list_documents(server)
server$close()
\dontshow{\}) # examplesIf}
}
