% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{print.amsync_server}
\alias{print.amsync_server}
\title{Print method for amsync_server}
\usage{
\method{print}{amsync_server}(x, ...)
}
\arguments{
\item{x}{An amsync_server object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns x.
}
\description{
Print method for amsync_server
}
\keyword{internal}
