% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{print.autosync_server}
\alias{print.autosync_server}
\title{Print method for autosync_server}
\usage{
\method{print}{autosync_server}(x, ...)
}
\arguments{
\item{x}{An autosync_server object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns x.
}
\description{
Print method for autosync_server
}
\keyword{internal}
