% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{print.sync_server}
\alias{print.sync_server}
\title{Print method for sync_server}
\usage{
\method{print}{sync_server}(x, ...)
}
\arguments{
\item{x}{A sync_server object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns x.
}
\description{
Print method for sync_server
}
\keyword{internal}
