% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stop_server}
\alias{stop_server}
\title{Stop the sync server}
\usage{
stop_server(server)
}
\arguments{
\item{server}{An amsync_server object.}
}
\value{
Invisibly returns the server object.
}
\description{
Stops a running sync server and cleans up resources.
}
