% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{list_documents}
\alias{list_documents}
\title{List all document IDs}
\usage{
list_documents(server)
}
\arguments{
\item{server}{An amsync_server object.}
}
\value{
Character vector of document IDs.
}
\description{
Returns the IDs of all documents currently loaded in the server.
}
